some edits with users

+++ backend/src/modules/users/services/user-follows.service.ts
@@ -1,11 +1,15 @@
-import { BadRequestException, Injectable } from "@nestjs/common";
-import { PrismaService } from "src/modules/prisma/prisma.service";
-import { HelpersService } from "src/modules/helpers/services/helpers.service";
-import { Follow, User } from "@prisma/client";
+import { BadRequestException, Injectable } from '@nestjs/common';
+import { PrismaService } from 'src/modules/prisma/prisma.service';
+import { HelpersService } from 'src/modules/helpers/services/helpers.service';
+import { User } from '@prisma/client';
+import { AuthRequest } from 'src/common/types';
 
 @Injectable()
 export class UserFollowsService {
-  constructor(private readonly prisma: PrismaService, private readonly helperService: HelpersService) {}
+  constructor(
+    private readonly prisma: PrismaService,
+    private readonly helperService: HelpersService,
+  ) {}
 
   /**
    * Find all followers of a user
@@ -32,12 +36,17 @@ export class UserFollowsService {
 
   /**
    * Follow a user
-   * @param userId - The ID of the user to follow
+   * @param userId - The ID of the user
    * @param followerId - The ID of the follower
    * @returns The follow object
    */
   async follow(userId: number, followerId: number) {
     try {
+
+      if (!userId) {
+        throw new BadRequestException('User not authenticated');
+      }
+
       await this.helperService.getIdOrThrow<User>('user', userId, 'User');
       await this.helperService.getIdOrThrow<User>('user', followerId, 'User');
 
@@ -57,10 +66,10 @@ export class UserFollowsService {
       if (follow) {
         throw new BadRequestException('You are already following this user');
       }
-  
+
       await this.prisma.follow.create({
         data: {
-          followerId,    
+          followerId,
           followingId: userId,
         },
       });
@@ -79,26 +88,50 @@ export class UserFollowsService {
       );
     }
   }
-  
+
   /**
    * Unfollow a user
-   * @param userId - The ID of the user to unfollow
+   * @param userId - The ID of the user
    * @param followerId - The ID of the follower
    * @returns The follow object
    */
   async unfollow(userId: number, followerId: number) {
     try {
+
+      await this.helperService.getIdOrThrow<User>('user', userId, 'User');
+      await this.helperService.getIdOrThrow<User>('user', followerId, 'User');
+      const follow = await this.prisma.follow.findUnique({
+        where: {
+          followerId_followingId: {
+            followerId,
+            followingId: userId,
+          },
+        },
+      });
+
+      if (!follow) {
+        throw new BadRequestException('You are not following this user');
+      }
+
       await this.prisma.follow.deleteMany({
         where: {
           followerId,
           followingId: userId,
         },
       });
+
+      return this.prisma.follow.findMany({
+        where: {
+          followingId: userId,
+        },
+        include: {
+          followedUser: true,
+        },
+      });
     } catch (error) {
       throw new BadRequestException(
-        `Error unfollowing user with ID ${userId}: ${error.message}`,
+        `Error unfollowing user with ID ${userId} and follower ID ${followerId}: ${error.message}`,
       );
     }
   }
-
 }
